fix: mobile client crash
